+++ src/ui/ui_text.c
@@ -68,7 +68,7 @@ static void update_dimensions(UI_Text *ui_text, const short source_id)
 {
   if (ui_text->area != source_id)
   {
-    int size[2];
+    short size[2];
 
     ui_text->area = source_id;
     UI_area_size(source_id, size);
